+++ ufl/algorithms/analysis.py
@@ -43,14 +43,6 @@ def unique_tuple(objects):
 
 # --- Utilities to extract information from an expression ---
 
-def __unused__extract_classes(a):
-    """Build a set of all unique Expr subclasses used in a.
-    The argument a can be a BaseForm, Integral or Expr."""
-    return set(o._ufl_class_
-               for e in iter_expressions(a)
-               for o in unique_pre_traversal(e))
-
-
 def extract_type(a, ufl_types):
     """Build a set of all objects found in a whose class is in ufl_types.
     The argument a can be a BaseForm, Integral or Expr."""
@@ -74,17 +66,6 @@ def extract_type(a, ufl_types):
         return set(o for e in iter_expressions(a)
                    for o in unique_pre_traversal(e)
                    if any(isinstance(o, t) for t in ufl_types))
-
-
-def has_type(a, ufl_type):
-    """Return if an object of class ufl_type can be found in a.
-    The argument a can be a BaseForm, Integral or Expr."""
-    if issubclass(ufl_type, Terminal):
-        # Optimization
-        traversal = traverse_unique_terminals
-    else:
-        traversal = unique_pre_traversal
-    return any(isinstance(o, ufl_type) for e in iter_expressions(a) for o in traversal(e))
 
 
 def has_exact_type(a, ufl_type):
