+++ ufl/operators.py
@@ -159,15 +159,6 @@ def inner(a, b):
     return Inner(a, b)
 
 
-# TODO: Something like this would be useful in some cases, but should
-# inner just support len(a.ufl_shape) != len(b.ufl_shape) instead?
-def _partial_inner(a, b):
-    "UFL operator: Take the partial inner product of a and b."
-    ar, br = len(a.ufl_shape), len(b.ufl_shape)
-    n = min(ar, br)
-    return contraction(a, list(range(n - ar, n - ar + n)), b, list(range(n)))
-
-
 def dot(a, b):
     "UFL operator: Take the dot product of *a* and *b*. This won't take the complex conjugate of the second argument."
     a = as_ufl(a)
@@ -175,30 +166,6 @@ def dot(a, b):
     if a.ufl_shape == () and b.ufl_shape == ():
         return a * b
     return Dot(a, b)
-
-
-def contraction(a, a_axes, b, b_axes):
-    "UFL operator: Take the contraction of a and b over given axes."
-    ai, bi = a_axes, b_axes
-    if len(ai) != len(bi):
-        raise ValueError("Contraction must be over the same number of axes.")
-    ash = a.ufl_shape
-    bsh = b.ufl_shape
-    aii = indices(len(a.ufl_shape))
-    bii = indices(len(b.ufl_shape))
-    cii = indices(len(ai))
-    shape = [None] * len(ai)
-    for i, j in enumerate(ai):
-        aii[j] = cii[i]
-        shape[i] = ash[j]
-    for i, j in enumerate(bi):
-        bii[j] = cii[i]
-        if shape[i] != bsh[j]:
-            raise ValueError("Shape mismatch in contraction.")
-    s = a[aii] * b[bii]
-    cii = set(cii)
-    ii = tuple(i for i in (aii + bii) if i not in cii)
-    return as_tensor(s, ii)
 
 
 def perp(v):
@@ -315,11 +282,6 @@ def Dx(f, *i):
     to spatial variable number *i*. Equivalent to ``f.dx(*i)``."""
     f = as_ufl(f)
     return f.dx(*i)
-
-
-def Dt(f):
-    "UFL operator: <Not implemented yet!> The partial derivative of *f* with respect to time."
-    raise NotImplementedError
 
 
 def Dn(f):
